+++ lndr-backend/lndr-backend.cabal
@@ -14,7 +14,12 @@ library
   hs-source-dirs:      src
   exposed-modules:     Docs
                      , EthInterface
+                     , Handler
+                     , Handler.Credit
+                     , Handler.Friend
+                     , Handler.Types
                      , Server
+                     , Types
   build-depends:       base >= 4.7 && < 5
                      , aeson
                      , base16-bytestring
@@ -44,17 +49,7 @@ executable lndr-server
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N
   build-depends:       base
                      , lndr-backend
-                     , aeson
-                     , either
-                     , http-types
-                     , lens
-                     , list-t
-                     , mtl
-                     , servant-server
-                     , text
-                     , transformers
                      , wai
                      , wai-logger
                      , warp
-                     , web3
   default-language:    Haskell2010
